+++ NEWS.md
@@ -4,6 +4,7 @@
 * Update variable lists.
 * Add a recid lookup for use with the `recid` filtering parameter.
 * Now depends on R >= 3.5
+* When using `get_anon_chi()` it will now use `phsmethods::chi_check` (if avaliable) and warn if a lot of CHIs look invalid.
 
 # slfhelper 0.6.0